+++ converter/internal/promtailconvert/promtailconvert.go
@@ -167,4 +167,5 @@ func appendScrapeConfig(
 	b.AppendJournalConfig()
 	b.AppendPushAPI()
 	b.AppendSyslogConfig()
+	b.AppendGCPLog()
 }